Updating Board.cs
Added comments to member variables and constructor.

Added test method to try to only have to print the things inside the board instead of the board.

Added code to print snake position as well as board dimentions to validate that the snake is spawning correctly..

Added code to spawn fruit in board.

Added code to check if the player ate the fruit.

Added code to check if snake hit the board.

Added code to return if the wall was hit and to reset the wall hit boolean variable.